refactored file factory into its own module
